+++ src/Osintgram.py
@@ -2,9 +2,13 @@ import datetime
 import json
 import sys
 import urllib
+import os
+import codecs
 
 from geopy.geocoders import Nominatim
 from instagram_private_api import Client as AppClient
+from instagram_private_api import ClientCookieExpiredError, ClientLoginRequiredError, ClientError
+
 from prettytable import PrettyTable
 
 from src import printcolors as pc
@@ -25,7 +29,7 @@ class Osintgram:
         u = self.__getUsername__()
         p = self.__getPassword__()
         print("\nAttempt to login...")
-        self.api = AppClient(auto_patch=True, authenticate=True, username=u, password=p)
+        self.login(u, p)
         self.setTarget(target)
         self.writeFile = is_file
         self.jsonDump = is_json
@@ -88,10 +92,11 @@ class Osintgram:
     def __printTargetBanner__(self):
         pc.printout("\nLogged as ", pc.GREEN)
         pc.printout(self.api.username, pc.CYAN)
-        pc.printout(" (" + str(self.api.authenticated_user_id) + ") ")
-        pc.printout("target: ", pc.GREEN)
+        pc.printout(". Target: ", pc.GREEN)
         pc.printout(str(self.target), pc.CYAN)
-        pc.printout(" (private: " + str(self.is_private) + ")")
+        pc.printout(" [" + str(self.target_id) + "] ")
+        if self.is_private:
+            pc.printout("[PRIVATE PROFILE]", pc.RED)
         print('\n')
 
     def change_target(self):
@@ -240,6 +245,7 @@ class Osintgram:
 
         for post in data:
             comments_counter += post['comment_count']
+            posts += 1
 
         if self.writeFile:
             file_name = "output/" + self.target + "_comments.txt"
@@ -488,6 +494,7 @@ class Osintgram:
 
         for post in data:
             like_counter += post['like_count']
+            posts += 1
 
         if self.writeFile:
             file_name = "output/" + self.target + "_likes.txt"
@@ -907,3 +914,57 @@ class Osintgram:
             pc.printout("\n")
 
         self.jsonDump = flag
+
+    def login(self, u, p):
+        try:
+            settings_file = "config/settings.json"
+            if not os.path.isfile(settings_file):
+                # settings file does not exist
+                print('Unable to find file: {0!s}'.format(settings_file))
+
+                # login new
+                self.api = AppClient(auto_patch=True, authenticate=True, username=u, password=p,
+                                     on_login=lambda x: self.onlogin_callback(x, settings_file))
+
+            else:
+                with open(settings_file) as file_data:
+                    cached_settings = json.load(file_data, object_hook=self.from_json)
+                #print('Reusing settings: {0!s}'.format(settings_file))
+
+                # reuse auth settings
+                self.api = AppClient(
+                    username=u, password=p,
+                    settings=cached_settings,
+                    on_login=lambda x: self.onlogin_callback(x, settings_file))
+
+        except (ClientCookieExpiredError, ClientLoginRequiredError) as e:
+            print('ClientCookieExpiredError/ClientLoginRequiredError: {0!s}'.format(e))
+
+            # Login expired
+            # Do relogin but use default ua, keys and such
+            self.api = AppClient(auto_patch=True, authenticate=True, username=u, password=p,
+                                 on_login=lambda x: self.onlogin_callback(x, settings_file))
+
+        except ClientError as e:
+            #pc.printout('ClientError {0!s} (Code: {1:d}, Response: {2!s})'.format(e.msg, e.code, e.error_response), pc.RED)
+            error = json.loads(e.error_response)
+            pc.printout(error['message'], pc.RED)
+            pc.printout("\n")
+            exit(9)
+
+    def to_json(self, python_object):
+        if isinstance(python_object, bytes):
+            return {'__class__': 'bytes',
+                    '__value__': codecs.encode(python_object, 'base64').decode()}
+        raise TypeError(repr(python_object) + ' is not JSON serializable')
+
+    def from_json(self, json_object):
+        if '__class__' in json_object and json_object['__class__'] == 'bytes':
+            return codecs.decode(json_object['__value__'].encode(), 'base64')
+        return json_object
+
+    def onlogin_callback(self, api, new_settings_file):
+        cache_settings = api.settings
+        with open(new_settings_file, 'w') as outfile:
+            json.dump(cache_settings, outfile, default=self.to_json)
+            #print('SAVED: {0!s}'.format(new_settings_file))